# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MCP Finance System API",
        "description": "API demo financiera construida en Laravel 12, protegida con Sanctum y consumida por un servidor MCP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8002",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión y obtener token Sanctum",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login correcto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|XXXXXXXXX"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Admin Demo"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/system/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lista de usuarios del sistema financiero",
                "operationId": "8133e2d1ee03ca1fad363c765e55e8e1",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Juan Pérez"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "juan@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/system/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Detalle de un usuario por ID",
                "operationId": "98aad933eb4eba3d95102eb3a4eb25e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/system/users/{id}/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Listar cuentas de un usuario",
                "operationId": "62ad5a0247e67242cbf7cc5abb2b7e18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de cuentas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Account"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/system/accounts/{accountId}/transactions": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Transacciones de una cuenta",
                "operationId": "034649db1df707183198b9636ffd314b",
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "description": "ID de la cuenta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de transacciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cuenta no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/system/users/{id}/financial-summary": {
            "get": {
                "tags": [
                    "Summary"
                ],
                "summary": "Resumen financiero de un usuario",
                "operationId": "5e936bc7205c6c8fb59be135aa3d0010",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resumen financiero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FinancialSummary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "example": "juan@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Account": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "account_number": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "account_type": {
                        "type": "string",
                        "example": "savings"
                    },
                    "currency": {
                        "type": "string",
                        "example": "COP"
                    },
                    "balance": {
                        "type": "number",
                        "format": "float",
                        "example": 1500000.75
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 100
                    },
                    "account_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "posted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-01T09:30:00Z"
                    },
                    "description": {
                        "type": "string",
                        "example": "Pago supermercado"
                    },
                    "category": {
                        "type": "string",
                        "example": "groceries"
                    },
                    "direction": {
                        "type": "string",
                        "example": "debit"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 250000
                    }
                },
                "type": "object"
            },
            "FinancialSummary": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "accounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Account"
                        }
                    },
                    "totals": {
                        "properties": {
                            "total_balance": {
                                "type": "number",
                                "format": "float",
                                "example": 3500000.5
                            },
                            "total_income": {
                                "type": "number",
                                "format": "float",
                                "example": 5000000
                            },
                            "total_expense": {
                                "type": "number",
                                "format": "float",
                                "example": 1500000.5
                            },
                            "period": {
                                "type": "string",
                                "example": "2025-01"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de Laravel Sanctum. Enviar como: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Users",
            "description": "Users"
        },
        {
            "name": "Accounts",
            "description": "Accounts"
        },
        {
            "name": "Summary",
            "description": "Summary"
        }
    ]
}